+++ expyriment/misc/geometry/__init__.py
@@ -14,5 +14,5 @@ __date__ = ''
 
 from ._geometry import coordinates2position, position2coordinates, position2coordinate
 from ._geometry import position2visual_angle, visual_angle2position
-from ._geometry import cartesian2polar, polar2cartesian
+from ._geometry import cartesian2polar, polar2cartesian, tuples2points
 from ._geometry import points_to_vertices, points2vertices, lines_intersect, XYPoint